DF-16088 Fix return value from GetMsSinceEpoch()
Former-commit-id: a826b1d9420eec81f4ed84ba8c830cb7748d930a

+++ Sources/Internal/Autotesting/AutotestingSystemLua.h
@@ -64,7 +64,7 @@ public:
     size_t GetUsedMemory() const;
 
     float32 GetTimeElapsed();
-    float32 GetMsSinceEpoch();
+    int32 GetMsSinceEpoch();
 
     // Test organization API
     void OnTestStart(const String& testName);
